+++ test/unit/braintree/client_token_gateway_spec.js
@@ -3,7 +3,7 @@
 let braintree = specHelper.braintree;
 
 describe("ClientTokenGateway", () =>
-  describe("generate", () =>
+  describe("generate", () => {
     it("returns an error when credit card options are supplied without a customer ID", function (done) {
       specHelper.defaultGateway.clientToken.generate(
         {
@@ -18,4 +18,19 @@ describe("ClientTokenGateway", () =>
           done();
         }
       );
-    })));
+    });
+
+    it("does not return an error when the domains option is supplied", function (done) {
+      specHelper.defaultGateway.clientToken.generate(
+        {
+          domains: ["example.com"],
+        },
+        function (err, response) {
+          assert.equal(err, null);
+          assert.typeOf(response, "object");
+          assert.typeOf(response.clientToken, "string");
+          done();
+        }
+      );
+    });
+  }));